Multiple textures, images are loaded with SOIL2.

+++ include/core/threads/thread_safe_container.h
@@ -44,7 +44,7 @@ public:
 		return true;
 	}
 
-	int size() {
+	size_t size() {
 		std::unique_lock<std::mutex> lock(mtx);
 		return queue.size();
 	}
